update: seperate ui and state for checkbox
